+++ avalon-modules/avalon-dataset/avalon-dataset-repository/src/main/java/io/github/lishangbu/avalon/dataset/repository/ItemFlingEffectRepository.java
@@ -13,5 +13,5 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface ItemFlingEffectRepository
-    extends ListCrudRepository<ItemFlingEffect, Long>,
-        ListPagingAndSortingRepository<ItemFlingEffect, Long> {}
+    extends ListCrudRepository<ItemFlingEffect, Integer>,
+        ListPagingAndSortingRepository<ItemFlingEffect, Integer> {}
