+++ avalon-modules/avalon-dataset/avalon-dataset-repository/src/main/java/io/github/lishangbu/avalon/dataset/repository/PokemonTypeRepository.java
@@ -13,5 +13,5 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface PokemonTypeRepository
-    extends ListCrudRepository<PokemonType, Long>,
-        ListPagingAndSortingRepository<PokemonType, Long> {}
+    extends ListCrudRepository<PokemonType, Integer>,
+        ListPagingAndSortingRepository<PokemonType, Integer> {}
